Create the PhpSettingsConfiguration middleware
